+++ Source/Rooms/empty.js
@@ -5,9 +5,13 @@ module.exports = new RoomTemplate("Empty Room", [
 ]).setDescription("This room is empty. Lucky you?")
 	.setElement("Untyped");
 
-module.exports.uiRows.push(new ActionRowBuilder().addComponents(
-	new ButtonBuilder().setCustomId("continue")
-		.setEmoji("👑")
-		.setLabel("Move on")
-		.setStyle(ButtonStyle.Secondary)
-))
+module.exports.buildUI = function (adventure) {
+	return [
+		new ActionRowBuilder().addComponents(
+			new ButtonBuilder().setCustomId("continue")
+				.setEmoji("👑")
+				.setLabel("Move on")
+				.setStyle(ButtonStyle.Secondary)
+		)
+	];
+}