+++ Source/Rooms/event-freegoldonfire.js
@@ -7,9 +7,13 @@ module.exports = new RoomTemplate("Free Gold?", [
 ]).setDescription("A large pile of gold sits quietly in the middle of the room, seemingly alone.")
 	.setElement("Fire");
 
-module.exports.uiRows.push(new ActionRowBuilder().addComponents(
-	new ButtonBuilder()
-		.setCustomId("getgoldonfire")
-		.setLabel("Would be a waste to leave it there [+300 gold]")
-		.setStyle(ButtonStyle.Danger)
-))
+module.exports.buildUI = function (adventure) {
+	return [
+		new ActionRowBuilder().addComponents(
+			new ButtonBuilder()
+				.setCustomId("getgoldonfire")
+				.setLabel("Would be a waste to leave it there [+300 gold]")
+				.setStyle(ButtonStyle.Danger)
+		)
+	];
+}