+++ Source/labyrinths/debugdungeon.js
@@ -167,7 +167,7 @@ module.exports = new Labyrinth("Debug Dungeon", "Untyped", 10, [10])
 		})
 	.setRooms(
 		{
-			"Event": ["Elemental Research", "Free Gold?", "Health Redistribution", "The Score Beggar", "Repair Kit, just hanging out", "Abandoned Forge", "Equipment Merchant", "Consumable Merchant", "Rest Site", "Treasure!"],
+			"Event": ["Twin Pedestals", "Elemental Research", "Free Gold?", "Health Redistribution", "The Score Beggar", "Repair Kit, just hanging out", "Abandoned Forge", "Equipment Merchant", "Consumable Merchant", "Rest Site", "Treasure!"],
 			"Battle": ["Hawk Fight", "Frog Fight", "Mechabee Fight", "Slime Fight", "Tortoise Fight"],
 			"Merchant": ["Equipment Merchant", "Consumable Merchant"],
 			"Rest Site": ["Rest Site"],